+++ packages/common/src/compiler/filtersCompiler.ts
@@ -1,4 +1,4 @@
-import moment from 'moment/moment';
+import moment from 'moment-timezone';
 import { SupportedDbtAdapter } from '../types/dbt';
 import {
     CustomFormatType,
@@ -139,6 +139,7 @@ export const renderDateFilterSql = (
     dimensionSql: string,
     filter: DateFilterRule,
     adapterType: SupportedDbtAdapter,
+    timezone: string,
     dateFormatter: (date: Date) => string = formatDate,
     startOfWeek: WeekDay | null | undefined = undefined,
 ): string => {
@@ -262,12 +263,16 @@ export const renderDateFilterSql = (
                 filter.settings?.unitOfTime || UnitOfTime.days;
             const fromDate = dateFormatter(
                 getMomentDateWithCustomStartOfWeek(startOfWeek)
+                    .tz(timezone)
                     .startOf(unitOfTime)
+                    .utc()
                     .toDate(),
             );
             const untilDate = dateFormatter(
                 getMomentDateWithCustomStartOfWeek(startOfWeek)
+                    .tz(timezone)
                     .endOf(unitOfTime)
+                    .utc()
                     .toDate(),
             );
             return `((${dimensionSql}) >= ${castValue(
@@ -319,6 +324,7 @@ export const renderTableCalculationFilterRuleSql = (
     escapeStringQuoteChar: string,
     adapterType: SupportedDbtAdapter,
     startOfWeek: WeekDay | null | undefined,
+    timezone: string = 'UTC',
 ): string => {
     if (!field) return '1=1';
 
@@ -340,6 +346,7 @@ export const renderTableCalculationFilterRuleSql = (
                 fieldSql,
                 filterRule,
                 adapterType,
+                timezone,
                 undefined,
                 startOfWeek,
             );
@@ -375,26 +382,26 @@ export const renderFilterRuleSql = (
     escapeStringQuoteChar: string,
     startOfWeek: WeekDay | null | undefined,
     adapterType: SupportedDbtAdapter,
-    timezone?: string, // TODO replacde with enum
+    timezone: string = 'UTC',
 ): string => {
     if (filterRule.disabled) {
         return `1=1`; // When filter is disabled, we want to return all rows
     }
-
-    const convertBigqueryTimezone = (originalFieldSql: string) => {
-        // On Bigquery we convert timestamps to the right timezone before adding the SQL filter
-        // Bigquery does not support set TIMEZONE in session like the rest of the warehouses
-        // and field.compiledSql is generated in compile time, so we need to patch it here
-        // Only timestamp type in Bigquery has timezone information.
-        if (timezone && adapterType === SupportedDbtAdapter.BIGQUERY) {
-            const timestampRegex = /TIMESTAMP_TRUNC\(([^,]+),/;
-            return originalFieldSql.replace(
-                timestampRegex,
-                `TIMESTAMP_TRUNC(DATE($1, '${timezone}'),`,
-            );
-        }
-        return originalFieldSql;
-    };
+    // Note: This function is WIP for the user timezone milestone: https://github.com/lightdash/lightdash/milestone/97
+    // const convertBigqueryTimezone = (originalFieldSql: string) => {
+    //     // On Bigquery we convert timestamps to the right timezone before adding the SQL filter
+    //     // Bigquery does not support set TIMEZONE in session like the rest of the warehouses
+    //     // and field.compiledSql is generated in compile time, so we need to patch it here
+    //     // Only timestamp type in Bigquery has timezone information.
+    //     if (timezone && adapterType === SupportedDbtAdapter.BIGQUERY) {
+    //         const timestampRegex = /TIMESTAMP_TRUNC\(([^,]+),/;
+    //         return originalFieldSql.replace(
+    //             timestampRegex,
+    //             `TIMESTAMP_TRUNC(DATE($1, '${timezone}'),`,
+    //         );
+    //     }
+    //     return originalFieldSql;
+    // };
     const fieldType = isCompiledCustomSqlDimension(field)
         ? field.dimensionType
         : field.type;
@@ -430,6 +437,7 @@ export const renderFilterRuleSql = (
                 fieldSql,
                 filterRule,
                 adapterType,
+                timezone,
                 undefined,
                 startOfWeek,
             );
@@ -437,9 +445,10 @@ export const renderFilterRuleSql = (
         case DimensionType.TIMESTAMP:
         case MetricType.TIMESTAMP: {
             return renderDateFilterSql(
-                convertBigqueryTimezone(fieldSql),
+                fieldSql,
                 filterRule,
                 adapterType,
+                timezone,
                 formatTimestampAsUTCWithNoTimezone,
                 startOfWeek,
             );