+++ packages/e2e/cypress/e2e/app/login.cy.ts
@@ -14,6 +14,7 @@ describe('Login', () => {
         cy.findByPlaceholderText('Your email address').type(
             SEED_ORG_1_ADMIN_EMAIL.email,
         );
+        cy.findByText('Continue').click();
         cy.findByPlaceholderText('Your password').type(
             SEED_ORG_1_ADMIN_PASSWORD.password,
         );
@@ -25,16 +26,15 @@ describe('Login', () => {
         cy.logout();
         cy.visit('/login');
         cy.findByPlaceholderText('Your email address').type('test-email');
-        cy.findByPlaceholderText('Your password').type('test-password');
-        cy.get('[data-cy="signin-button"]').click();
+        cy.findByText('Continue').click();
         cy.findByText('Email address is not valid').should('be.visible');
         cy.findByPlaceholderText('Your email address').type('test@emaill.com');
+        cy.findByText('Continue').click();
         cy.findByPlaceholderText('Your password').type('test-password');
         cy.get('[data-cy="signin-button"]').click();
         cy.findByText('Email and password not recognized').should('be.visible');
         cy.findByPlaceholderText('Your email address').type('test@email.com ');
-        cy.findByPlaceholderText('Your password').type('test-password');
-        cy.get('[data-cy="signin-button"]').click();
+        cy.findByText('Continue').click();
         cy.findByText('Email address must not contain whitespaces').should(
             'be.visible',
         );